# server/internal/api/api.gn.go
//go:build go1.22

// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.7.0 DO NOT EDIT.
package api

import (
	"fmt"
	"net/http"

	"github.com/oapi-codegen/runtime"
	openapi_types "github.com/oapi-codegen/runtime/types"
)

// Book defines model for Book.
type Book struct {
	AuthorId      int      `json:"author_id"`
	AuthorName    string   `json:"author_name"`
	CoverImage    string   `json:"cover_image"`
	Description   string   `json:"description"`
	Id            int      `json:"id"`
	Tags          TagArray `json:"tags"`
	Title         string   `json:"title"`
	YearOfRelease int      `json:"year_of_release"`
}

// BookArray defines model for BookArray.
type BookArray = []Book

// BookPost defines model for BookPost.
type BookPost struct {
	AuthorName    string   `json:"author_name"`
	Description   string   `json:"description"`
	Tags          []string `json:"tags"`
	Title         string   `json:"title"`
	YearOfRelease int      `json:"year_of_release"`
}

// LoginCredentials defines model for LoginCredentials.
type LoginCredentials struct {
	Email    string `json:"email"`
	Password string `json:"password"`
}

// Rating defines model for Rating.
type Rating struct {
	Name  string  `json:"name"`
	Value float32 `json:"value"`
}

// RatingRule defines model for RatingRule.
type RatingRule struct {
	Name         string  `json:"name"`
	ValueCeiling float32 `json:"value_ceiling"`
}

// Review defines model for Review.
type Review struct {
	Ratings  []Rating `json:"ratings"`
	UserId   int      `json:"user_id"`
	Username string   `json:"username"`
}

// ReviewPost defines model for ReviewPost.
type ReviewPost struct {
	Ratings []Rating `json:"ratings"`
}

// ReviewRules defines model for ReviewRules.
type ReviewRules struct {
	RatingRules []RatingRule `json:"rating_rules"`
}

// SignupCredentials defines model for SignupCredentials.
type SignupCredentials struct {
	Email    string `json:"email"`
	Password string `json:"password"`
	Username string `json:"username"`
}

// Tag defines model for Tag.
type Tag struct {
	Id   int    `json:"id"`
	Name string `json:"name"`
}

// TagArray defines model for TagArray.
type TagArray = []Tag

// PostBooksMultipartBody defines parameters for PostBooks.
type PostBooksMultipartBody struct {
	Data  BookPost           `json:"data"`
	Image openapi_types.File `json:"image"`
}

// PostBooksMultipartRequestBody defines body for PostBooks for multipart/form-data ContentType.
type PostBooksMultipartRequestBody PostBooksMultipartBody

// PostBooksIdReviewsJSONRequestBody defines body for PostBooksIdReviews for application/json ContentType.
type PostBooksIdReviewsJSONRequestBody = ReviewPost

// PostLoginJSONRequestBody defines body for PostLogin for application/json ContentType.
type PostLoginJSONRequestBody = LoginCredentials

// PostSignupJSONRequestBody defines body for PostSignup for application/json ContentType.
type PostSignupJSONRequestBody = SignupCredentials

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// Returns a list of all books
	// (GET /books)
	GetBooks(w http.ResponseWriter, r *http.Request)
	// Creates new book
	// (POST /books)
	PostBooks(w http.ResponseWriter, r *http.Request)
	// Returns a list of reviews
	// (GET /books/{id}/reviews)
	GetBooksIdReviews(w http.ResponseWriter, r *http.Request, id int)
	// Creates new review
	// (POST /books/{id}/reviews)
	PostBooksIdReviews(w http.ResponseWriter, r *http.Request, id int)
	// Logs user in if the password is correct
	// (POST /login)
	PostLogin(w http.ResponseWriter, r *http.Request)
	// Returns rules for a review
	// (GET /review-rules)
	GetReviewRules(w http.ResponseWriter, r *http.Request)
	// Signs user up if the email isn't taken
	// (POST /signup)
	PostSignup(w http.ResponseWriter, r *http.Request)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandlerFunc   func(w http.ResponseWriter, r *http.Request, err error)
}

type MiddlewareFunc func(http.Handler) http.Handler

// GetBooks operation middleware
func (siw *ServerInterfaceWrapper) GetBooks(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetBooks(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// PostBooks operation middleware
func (siw *ServerInterfaceWrapper) PostBooks(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.PostBooks(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetBooksIdReviews operation middleware
func (siw *ServerInterfaceWrapper) GetBooksIdReviews(w http.ResponseWriter, r *http.Request) {

	var err error
	_ = err

	// ------------- Path parameter "id" -------------
	var id int

	err = runtime.BindStyledParameterWithOptions("simple", "id", r.PathValue("id"), &id, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "integer", Format: ""})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "id", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetBooksIdReviews(w, r, id)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// PostBooksIdReviews operation middleware
func (siw *ServerInterfaceWrapper) PostBooksIdReviews(w http.ResponseWriter, r *http.Request) {

	var err error
	_ = err

	// ------------- Path parameter "id" -------------
	var id int

	err = runtime.BindStyledParameterWithOptions("simple", "id", r.PathValue("id"), &id, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "integer", Format: ""})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "id", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.PostBooksIdReviews(w, r, id)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// PostLogin operation middleware
func (siw *ServerInterfaceWrapper) PostLogin(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.PostLogin(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetReviewRules operation middleware
func (siw *ServerInterfaceWrapper) GetReviewRules(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetReviewRules(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// PostSignup operation middleware
func (siw *ServerInterfaceWrapper) PostSignup(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.PostSignup(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

type UnescapedCookieParamError struct {
	ParamName string
	Err       error
}

func (e *UnescapedCookieParamError) Error() string {
	return fmt.Sprintf("error unescaping cookie parameter '%s'", e.ParamName)
}

func (e *UnescapedCookieParamError) Unwrap() error {
	return e.Err
}

type UnmarshalingParamError struct {
	ParamName string
	Err       error
}

func (e *UnmarshalingParamError) Error() string {
	return fmt.Sprintf("Error unmarshaling parameter %s as JSON: %s", e.ParamName, e.Err.Error())
}

func (e *UnmarshalingParamError) Unwrap() error {
	return e.Err
}

type RequiredParamError struct {
	ParamName string
}

func (e *RequiredParamError) Error() string {
	return fmt.Sprintf("Query argument %s is required, but not found", e.ParamName)
}

type RequiredHeaderError struct {
	ParamName string
	Err       error
}

func (e *RequiredHeaderError) Error() string {
	return fmt.Sprintf("Header parameter %s is required, but not found", e.ParamName)
}

func (e *RequiredHeaderError) Unwrap() error {
	return e.Err
}

type InvalidParamFormatError struct {
	ParamName string
	Err       error
}

func (e *InvalidParamFormatError) Error() string {
	return fmt.Sprintf("Invalid format for parameter %s: %s", e.ParamName, e.Err.Error())
}

func (e *InvalidParamFormatError) Unwrap() error {
	return e.Err
}

type TooManyValuesForParamError struct {
	ParamName string
	Count     int
}

func (e *TooManyValuesForParamError) Error() string {
	return fmt.Sprintf("Expected one value for %s, got %d", e.ParamName, e.Count)
}

// Handler creates http.Handler with routing matching OpenAPI spec.
func Handler(si ServerInterface) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{})
}

// ServeMux is an abstraction of [http.ServeMux].
type ServeMux interface {
	HandleFunc(pattern string, handler func(http.ResponseWriter, *http.Request))
	http.Handler
}

type StdHTTPServerOptions struct {
	BaseURL          string
	BaseRouter       ServeMux
	Middlewares      []MiddlewareFunc
	ErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

// HandlerFromMux creates http.Handler with routing matching OpenAPI spec based on the provided mux.
func HandlerFromMux(si ServerInterface, m ServeMux) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{
		BaseRouter: m,
	})
}

func HandlerFromMuxWithBaseURL(si ServerInterface, m ServeMux, baseURL string) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{
		BaseURL:    baseURL,
		BaseRouter: m,
	})
}

// HandlerWithOptions creates http.Handler with additional options
func HandlerWithOptions(si ServerInterface, options StdHTTPServerOptions) http.Handler {
	m := options.BaseRouter

	if m == nil {
		m = http.NewServeMux()
	}
	if options.ErrorHandlerFunc == nil {
		options.ErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		}
	}

	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandlerFunc:   options.ErrorHandlerFunc,
	}

	m.HandleFunc(http.MethodGet+" "+options.BaseURL+"/books", wrapper.GetBooks)
	m.HandleFunc(http.MethodPost+" "+options.BaseURL+"/books", wrapper.PostBooks)
	m.HandleFunc(http.MethodGet+" "+options.BaseURL+"/books/{id}/reviews", wrapper.GetBooksIdReviews)
	m.HandleFunc(http.MethodPost+" "+options.BaseURL+"/books/{id}/reviews", wrapper.PostBooksIdReviews)
	m.HandleFunc(http.MethodPost+" "+options.BaseURL+"/login", wrapper.PostLogin)
	m.HandleFunc(http.MethodGet+" "+options.BaseURL+"/review-rules", wrapper.GetReviewRules)
	m.HandleFunc(http.MethodPost+" "+options.BaseURL+"/signup", wrapper.PostSignup)

	return m
}